Evol pour Alpha

+++ vertigo-orchestra-webapi/src/main/javagen/io/vertigo/orchestra/webapi/dao/uiexecutions/UiexecutionsPAO.java
@@ -56,11 +56,11 @@ public final class UiexecutionsPAO implements StoreServices {
 	 * Execute la tache TK_GET_EXECUTIONS_BY_PROCESS_NAME.
 	 * @param name String 
 	 * @param status String 
-	 * @param limit Long 
-	 * @param offset Long 
+	 * @param limit Integer 
+	 * @param offset Integer 
 	 * @return io.vertigo.dynamo.domain.model.DtList<io.vertigo.orchestra.webapi.domain.uiexecutions.OProcessExecutionUi> dtcOProcessExecutionUi
 	*/
-	public io.vertigo.dynamo.domain.model.DtList<io.vertigo.orchestra.webapi.domain.uiexecutions.OProcessExecutionUi> getExecutionsByProcessName(final String name, final String status, final Long limit, final Long offset) {
+	public io.vertigo.dynamo.domain.model.DtList<io.vertigo.orchestra.webapi.domain.uiexecutions.OProcessExecutionUi> getExecutionsByProcessName(final String name, final String status, final Integer limit, final Integer offset) {
 		final Task task = createTaskBuilder("TK_GET_EXECUTIONS_BY_PROCESS_NAME")
 				.addValue("NAME", name)
 				.addValue("STATUS", status)
